+++ Example/Sources/View Controllers/AdvancedExampleViewController.swift
@@ -26,6 +26,7 @@ import UIKit
 import MapKit
 import MessageKit
 import InputBarAccessoryView
+import PINRemoteImage
 
 final class AdvancedExampleViewController: ChatViewController {
         
@@ -369,6 +370,14 @@ extension AdvancedExampleViewController: MessagesDisplayDelegate {
         button.isUserInteractionEnabled = false // respond to accessoryView tap through `MessageCellDelegate`
         accessoryView.layer.cornerRadius = accessoryView.frame.height / 2
         accessoryView.backgroundColor = UIColor.primaryColor.withAlphaComponent(0.3)
+    }
+
+    func configureMediaMessageImageView(_ imageView: UIImageView, for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) {
+        if case MessageKind.photo(let media) = message.kind, let imageURL = media.url {
+            imageView.pin_setImage(from: imageURL)
+        } else {
+            imageView.pin_cancelImageDownload()
+        }
     }
     
     // MARK: - Location Messages