+++ Sources/Views/MessagesCollectionView.swift
@@ -57,7 +57,7 @@ open class MessagesCollectionView: UICollectionView {
 
     public override init(frame: CGRect, collectionViewLayout layout: UICollectionViewLayout) {
         super.init(frame: frame, collectionViewLayout: layout)
-        backgroundColor = .backgroundColor
+        backgroundColor = .collectionViewBackground
         registerReusableViews()
         setupGestureRecognizers()
     }